application/backend/requirements.txt
asgiref>=3.8.1
Django>=5.2.6
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
sqlparse>=0.5.2
gunicorn>=23.0.0
mysqlclient>=2.2.7
pillow>=11.3.0
django-filter>=25.2
django-cors-headers>=4.9.0
